+++ src/app/mega-menu-header/header/mega-menu/index.tsx
@@ -1,5 +1,5 @@
-import { Menu3 } from '@/app/mega-menu-header/header/nav-menu/menu3';
-import { Menu4 } from '@/app/mega-menu-header/header/nav-menu/menu4';
+import { Menu3 } from '@/app/mega-menu-header/header/mega-menu/menu3';
+import { Menu4 } from '@/app/mega-menu-header/header/mega-menu/menu4';
 import styles from './styles.module.css';
 
 type Props = {
@@ -7,7 +7,7 @@ type Props = {
   setIsHovering: (hoverState: string | null) => void;
 };
 
-export const NavMenu = ({ isHovering, setIsHovering }: Props) => {
+export const MegaMenu = ({ isHovering, setIsHovering }: Props) => {
   if (!isHovering) return null;
 
   return (
@@ -16,7 +16,7 @@ export const NavMenu = ({ isHovering, setIsHovering }: Props) => {
       onMouseEnter={() => setIsHovering(isHovering)}
       onMouseLeave={() => setIsHovering(null)}
     >
-      {isHovering === 'navMenu3' ? <Menu3 /> : <Menu4 />}
+      {isHovering === 'megaMenu3' ? <Menu3 /> : <Menu4 />}
     </nav>
   );
 };
