+++ src/app/mega-menu-header/header/mega-menu/list/index.tsx
@@ -1,4 +1,4 @@
-import styles from '@/app/mega-menu-header/header/nav-menu/menu3/styles.module.css';
+import styles from '@/app/mega-menu-header/header/mega-menu/menu3/styles.module.css';
 
 type Props = {
   items: string[];
