+++ src/app/mega-menu-header/header/mega-menu/menu4/faq-list/index.tsx
@@ -1,6 +1,6 @@
-import { List } from '@/app/mega-menu-header/header/nav-menu/list';
-import { faq1, faq2 } from '@/app/mega-menu-header/header/nav-menu/menu4/data';
-import styles from '@/app/mega-menu-header/header/nav-menu/menu4/styles.module.css';
+import { List } from '@/app/mega-menu-header/header/mega-menu/list';
+import { faq1, faq2 } from '@/app/mega-menu-header/header/mega-menu/menu4/data';
+import styles from '@/app/mega-menu-header/header/mega-menu/menu4/styles.module.css';
 
 export const FaqList = () => (
   <div className={`${styles.content} ${styles.contentWithBorder}`}>
